Add GestureDelegate
temp back up:
1. comment gesture register in any other UV
2. gestureDelegate is for better managing gesture manipulation of each UV
